pyproject.toml
[tool.poetry]
name = "pic-annotator"
version = "0.1.0"
description = ""
authors = ["Arthur2479 <arthurjacobs1004@gmail.com>"]
readme = "README.md"
packages = [{include = "pic_annotator"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
